Trying to modularize. auto-updating

+++ octave/arpack/build.sh
@@ -1,8 +1,10 @@
 #!/bin/bash
 NAME="arpack"
 PRGNAM="arpack-ng"
-VERSION=3.9.1
-wget -c https://github.com/opencollab/arpack-ng/archive/$VERSION.tar.gz -O $NAME-$VERSION.tar.gz
+VERSION=$(wget -cqO- https://github.com/opencollab/arpack-ng/tags | grep ".tar.gz" | head -n 1 | cut -d '"' -f 4 | cut -d '/' -f 7 | sed 's/.tar.gz//g')
+if ! [[ -f "$NAME-$VERSION.tar.xz" ]]; then
+	wget -c https://github.com/opencollab/arpack-ng/archive/$VERSION.tar.gz -O $NAME-$VERSION.tar.gz
+fi
 rm -rf $PRGNAM-$VERSION
 tar xf $NAME-$VERSION.tar.gz
 cd $PRGNAM-$VERSION
